Refactor email notification setting tests
The `send_email_notification_of_message?` flag comes from the settings in the
User model and it can be set in two ways, either setting as 'true' or 'false'
(as a string) directly or using the `save_settings!` method.

The traits used for these tests are removed to make it clearer how the tests
are set up.