Fix rubocop offences in #injection_errors tests
Fixes a number of issues with tests for #injection_errors in
base_claim_presenter_spec. This mostly involves the way that things are
stubbed and mocked in order to test the `InjectionAttemptErrorable`
module. The final product could almost certainly still be improved on,
but I think this is an improvement on what exists at present.