Increase Sidekiq logging level for tests
Sidekiq defaults to logging INFO for all environments. Recently (since
moving to Sidekiq v7?) this has resulted in connection info being output
when running tests.

This has no impact on tests but is visually noisy and a little distracting.

This commit updates the Sidekiq initializer to increase the logging level
for the test environment to WARN so that only more important information
is logged when running tests.