Improve segregation of tests
base_claim_spec has variables declared which are present across all
examples, but only used in a small number. This commit adds describe
blocks to fence off tests relating to instantiation of the class and
fee scheme-related methods. This allows us to be more precise with test
setup and only create variables where they are needed.