CFP-179 Zeitwerk: Fix ManagementInformation namespace
The file location of several classes in the ManagementInformation namespace
does not match the module declarations in the code. This hasn't been a
problem until now but Zeitwerk is particular about this.

This commit moves classes into the correct namespaces which makes
Zeitwerk happy.

It also changes the file structure of the related `quries` specs for
consistency.