CTSKF-241 Zeitwerk: Fix Extensions namespace
Based on file structure, all files in the lib/extensions directory
belong in the Extensions namespace, but aren't. This hasn't been a
problem until now but Zeitwerk is particular about this.

This commit moves all classes in the the lib/extensions directory into
the Extensions module and updates occurences in code, which makes
Zeitwerk happy.