Fix acronym inflection in migration classes
Previous commits changed the inflection of certain acronyms. The rails
zeitwerk check did not pick up the fact that those acronyms are used in
migration classes, however the fact that they are present in camel case
form (instead of upper case) is now causing database migrations to fail.

This commit updates the acronyms in those classes.