Add Response to the globals list
For some reason Response is not being recognised even though it is part of
Javascript (https://caniuse.com/?search=Response). This is possibly due to out
of date dependencies preventing ES6 coding standards being checkd.

It is added to the list of globals to allow `yarn standard` to pass. It may be
possible to remove this after refactoring.