Improve fetching of claims for fee scheme
Fetching the list of claims for a fee scheme is slow as it involves an
individual database look-up per claim. This is because the fee scheme of a
claim is not stored in the database.

To improve this the list of claims for a fee scheme is found as follows;

* Fetch the list of all representation order with dates within the range of the
  fee scheme. Pre-fetching is included for;
  * The defendants of these representation orders
  * The claims of these defendants
  * All defendants for these claims
  * All representation orders for these defendants
* This list of representation orders is converted into a unique list of claims
* The list of claims is filtered based on the date range for the representation
  order date and the main hearing date, as appropriate..