CTSKF-433 Fix clashing allocation table text
On the allocations table on `case_workers/admin/allocations`, when the
table is loading there is a clash of text. The table attempts to display
a message indicating that the page is loading and a message to indicate
that there are no records found, in the same place, which also clashes
with the table header. It looks ugly and is impossible to read.

This occurs because the datatable has not been properly configured in
`Modules.AllocationDataTable.js`. The loading text is held in the
`processing` language element, when it should be held as `loadingRecords`.

By changing this, only one of the messages is displayed at a time -
`loadingRecords` when the table is loading and `zeroRecords` when no
records are found.

After fixing this there is still a small issue with the processing widget
overlapping with the table header. This can be fixed by adding a top
margin to it in `app/webpack/stylesheets/components/_table.scss`.