CTSKF-445 Randomise jasmine tests
The jasmine tests for javascript code are not currently randomised. It
is desirable for tests to run in a random order as each test should be
isolated and independent of other tests.

This commit enables the `random` environment variable in `spec/support/jasmine-browser.json`.
This causes tests to start failing (highlighting why this is a problem).
These failures will be fixed in subsequent commits.