CTSKF-445 Fix teardown/setup of Allocation jasmine tests
When the jasmine test suite is run in a randomised order, failures
intermittently occur in tests in `spec/javascripts/Modules.AllocationDataTable_spec.js`.

There are two reasons for this:

1) In order for the tests to pass, `module.init()` needs to be called
before tests in the `...setAjaxURL` describe block are run, rather than
just relying on it having been called in previous tests. This commit adds
that call in a beforeEach block.

2) Errors also occur because a test in `Modules.AllocationScheme_spec.js`
changes the filtered scheme from AGFS to a dummy test value ('input-value').
When this test runs before the tests in `Modules.AllocationDataTable_spec.js`,
this change in state leaks through where tests expect the scheme to be AGFS,
leading to a failed test.

This commit fixes this in two ways:
* it adds teardown code to the test in `Modules.AllocationScheme_spec.js` to
reset the filter after the test has completed;
* it adds a call to `module.init()` in `Modules.AllocationDataTable_spec.js`
to ensure that the page is correctly set up before running the test.

This commit also DRYs up the code a little by using already-defined
`consts` instead of full module definitions.