CTSKF-445 Fix teardown of helpers-sidebar_spec.js
When running jasmine tests in a randomised order, tests in
`spec/javascripts/helpers-sidebar_spec.js` intermittently fail. This is
because some tests in that file are not torn down correctly and so affect
the running of other tests.

This commit fixes the problem by ensuring that the `.js-block` element
is reset after each test has run.