CTSKF-445 Fix Modules.OffenceSearchInput_spec.js
* Moves `$('body .mod-search-input').remove()` to an afterEach function,
instead of calling it as the first step of the beforeEach function. This
previous approach meant that the last test in this file was not torn
down correctly and caused subsequent tests to fail.

* Fixes the test defined in `describe('...trackUserInput'...` which
fails because `module.$input.val('mudr')` is called before module has
been initialised. This is done by calling `module.init()` before
`module.$input.val('mudr')`.