CTSKF-435 Fix setup/teardown of modules-DuplicateExpenseCtrl_spec.js
When running jasmine tests in a randomised order, tests in
`spec/javascripts/modules-DuplicateExpenseCtrl_spec.js` intermittently
fail. This is because some tests in that file are not setup and torn down
correctly and so affect the running of other tests.

This commit fixes the problem by reinstating a `$('body').append(domFixture);`
step which appears to have been erroneously commented out.

It also removes the `domFixture.empty()` from the `beforeEach` block
and changes the `afterAll` function (which already calls that method
to `afterEach`, as it makes more sense to tear down the previous test
after it has run, rather than at the start of the next test.