Consolidate report generation tasks
These 6 tasks are all identical except the report type so the common code is
created as a single task that takes a parament to determine which report is
generatied.

On the 'metrics' tab of the Sidekiq dashboard all these report tasks will
appear as the same. If it is desired to see these separately it may be
necessary to create multiple classes that inherit from a single base class.