Remove or_one method calls
Removing the NilExtension class has caused a failure when an API call
is made with a nil `actual_trial_length`, as this was relying on the
`or_one` method defined in the extension.

This could be resolved by reinstating the NilExtension class, but that
monkeypatching is unclear and we would probably be better off without it.

Instead this commit replaces the call to `or_one` in
app/interfaces/api/entities/cclf/base_claim.rb with the `||` operator
to return either the submitted trial length or 1 if it is nil.

For consistency this commit also removes the other place where `or_one`
is defined, `lib/extensions/integer_extension.rb`, and replaces any calls
to it with the `||` operator.