Remove monkey patching of Nil and Integer classes
Deletes the IntegerExtension and NilExtension classes and fixes handling
of nil/zero trial lengths in CCR/CCLF API claims.

Removing this monkey patching means that an alternative is needed to the
`or_one` methods that they defined. This is done by adding a method to
`app/interfaces/api/entities/base_entity.rb` which replicates the
functionality from the `or_one` method from the Integer class.

This method is called in `app/interfaces/api/entities/cclf/base_claim.rb`
and `app/interfaces/api/entities/ccr/base_claim.rb`, both of which
inherit from `base_entity.rb` and handles situations where trial lengths
are nil or zero, both of which are valid conditions when creating a claim
in CCCD but will cause the claim to fail to inject into CCR and CCLF.

For example:

1) An AGFS Final claim with a case type of Committal for Sentence in CCCD
does not require a user to enter dates for `estimated_trial_length`,
`actual_trial_length`, `retrial_estimated_length` or `retrial_actual_length`,
and these default to 0 in the database, however CCR will reject the claim.

2) An AGFS Final claim with a case type of Trial (questionably?) allows
a user to enter an `estimated_trial_length` of 0. Again CCR will reject
the claim.

3) An LGFS Final claim with a case type of Guilty Plea allows the user
to leave the `actual_trial_length` field blank. CCLF will reject the
claim.