Increase poll message count for production
The `SETTINGS__AWS__POLL_MESSAGE_COUNT` setting determines how many messages
are taken from the SQS queue for injection responses with each poll. Throughout
the day this queue builds up, although never to an unmanageable level. By
increasing the number of messages taken in production we should be able to
silence the occasional alerts that the queue is getting too large.