CTSKF-591 Add defendant dob to database anonymisation
The task used to create an anonymised database dump anonymises all
potentially sensitive data with the exception of defendants' dates of
birth.

This adds the `defendant.date_of_birth` field to the anonymiser, using
`Faker::Date.birthday` to generate a new value. This defaults to a
date of birth between 18 and 65, which will always be valid in CCCD.