CTSKF-591 Move deletion of old dumps in dump.rake
On completion of the anonymised database dump task, the user is given
the option of deleting all but the latest dump file. This currently
fails as the deletion is run from the developer's local machine, which
now cannot access the s3 bucket. Deletion tasks must be run from a
kubernetes pod.

This commit removes the prompt from the rake task that creates the
database dump and instead adds it to the task the copies the dump from
the s3 bucket. This must also now be run from a kubernetes pod, and so
prompting at this point means the deletion task will always now succeed.