Don't use Capybara.default_max_wait_time for sleeping
`Capybara.default_max_wait_time` is intended as a threshold for Capybara to
use when waiting for an Ajax request to complete:

https://github.com/teamcapybara/capybara/blob/master/README.md#asynchronous-javascript-ajax-and-friends

It is currently configured as 10 seconds in CCCD, but when used correctly
will return as soon as the request is complete/expected element is displayed
(ie it acts as a maximum time before erroring).

In one step definition, `default_max_wait_time` is being used to set a
the duration of a `sleep`. This means that all tests that use that step
will pause for 10 seconds, even though it is likely that the page will
be ready long before then.

In an attempt to speed up the test suite, this commit sets that sleep to
1 second, which does not seem to cause any failures.