Use wait instead of sleep in cucumber tests
Replaces a number of `sleep` calls in feature test step definitions with
`using_wait_time` blocks. `using_wait_time` returns early if the assertion
is successful, so using this approach should be quicker than sleeping
for an absolute amount of time.

In a couple of places where a `using_wait_time` block isn't suitable,
the `sleep` statement has been removed as an experiment to see if there
is any impact.