Reduce offence seed data for cucumber tests
When the cucumber test suite starts, a large volume of production seed
data is loaded into the database. This includes a full list of offences
for nine fee schemes, however only ten of those offences are used in
tests.

In an attempt to speed up the running of the test suite, this amends
`db/seeds/scheme_11.rb` to load a bespoke test-only version of the
offence data for AGFS fee scheme 11, containing only those ten offences.
This set of data is copied for subsequent AGFS fee schemes so the total
number of records being created is reduced from ~6,500 to 50.

* We could consider doing something similar for other seed data.
* We could consider not seeding the database at all and use factories to
generate required data instead.