Remove jasmine-jquery code
Remove jasmine-jquery npm package code that was hard coded in.

This is because the package is no longer being maintained and
it is already introducing breaking changes.

Also, it helps us move towards our long term aim of removing jQuery
as it's a large js package that is reducing performance and is no longer
needed.