Add 2 new fields to prisoner profile page
This commit adds 2 new fields to the prisoner profile page: post recall release date and license end date. This is to allow POMs to see the release dates for their recalls and other cases without having to go to c-nomis.
In order to display the most relevant date for post recall release date a method, recall_release_date, was created inside the OffenderBase class. This method handles the logic in deciding the most appropriate date to display in the offence_info view - and eventually displaying on the Prisoner Profile Page. It chooses between three dates: post_recall_release_override_date, post_recall_release_date, actual_parole_date.

For license end date, no logic is needed so it is simply added directly in the view.

This commit also includes a lot of robocop changes that are not specifically part of the ticket but part of a new implementation to enforce coding standards across the code base.
It’s preference being  what dates are present and which dates are earliest.