CTSKF-650 Fix API ActiveRecord::RecordNotFound messages
Rails 7 has changed the format of ActiveRecord::RecordNotFound error
messages returned by the `Model.find_by!` method:
https://github.com/rails/rails/pull/42454

In CCCD this means that where the message `Couldn't find DisbursementType`
was previously returned to API clients, this is now formatted as
`Couldn't find DisbursementType with [WHERE "disbursement_types"."unique_code" = $1]`.

Ideally SQL should not be included in messages sent to clients. In
addition to simply not looking very good this potentially represents a
security issue, albeit a very minor one, as it unnecessarily reveals
information about the structure of the database.

This commit fixes it by rescuing ActiveRecord::RecordNotFound errors in
the two places where they can be thrown (app/models/disbursement.rb and
app/models/expense.rb), and then re-raising with an error message
specified in a translation file.