Fix capture_output method
The capture_output method exists to allow items that are not strings to be used
by capture in view helpers. It converts the data using to_s. However, with
Rails 7 to_fs needs to be used instead for DateTime and some other data types.
If DateTime#to_s is used then the default formatting is used and the stting in
config/locales/en.yml is ignored.