Add request tests for correct order of rep orders
The test in spec/validators/representation_order_validator_spec.rb for the
correct ordering of representation order dates is failing in Rails 7+ due to
the force_validation field, as set in the set setup, not persisting during the
test. To ensure that the same functionality is still covered in testing new
requests tests have been created.