Remove deprecated govuk-button--disabled
See https://github.com/alphagov/govuk-frontend/issues/2271

The only place where this is used is the download button for the provisional
assessment by date report on the MI page. The button is (and was) only
appearing as disabled but could still be pressed as it is a link and not a
button.