CTSKF-483 Replace shell-spinner with tty-spinner
The shell-spinner gem is used to provide visual feedback that rake tasks
and scripts are running. shell-spinner is not actively supported and
uses the colorize gem which we would like to remove from CCCD.

This replaces shell-spinner with the tty-spinner gem, which has
comparable functionality, is more widely used and supported, and does
not depend on colorize.