CTSKF-483 Fix broken dump.rake task
Upgrading to rails 7 has caused the task to generate an anonymised
database dump to start to fail.

This has two causes:

1. The method `ActiveRecord::Base.connection_config` has been removed. This
replaces it with `ActiveRecord::Base.connection_db_config.configuration_hash`
to allow the tasks to run correctly.

2. The Arel method `Table.compile_insert(data)` has been removed. This replaces
the call to it in `dump_file_writer` with the recommended
`Arel::InsertManager.new(table).insert(data)`.