CTSKF-484 Remove colorize gem
CCCD uses the `colorize` gem to display coloured text in the terminal
(for example when running rake tasks and scripts).

`colorize` is released under the GPL licence, which has implications
for the licensing of CCCD.

This commit removes `colorize` and the single place where it is
referenced (lib/tasks/brakeman.rake). CCCD already uses the `awesome_print`
gem which provides coloured output so there is no requirement to add
a replacement.