Index fees table on uuid to fix slow query
Sentry is reporting that the most time-consuming query in CCCD is

`SELECT "fees".* FROM "fees" WHERE "fees"."uuid" = $1 ORDER BY "fees"."id" ASC LIMIT $2`

which takes on average ~400ms to complete.

This query originates in the following line in `app/interfaces/api/v1/external_users/date_attended.rb`:

`::Fee::BaseFee.find_by(uuid: params[:attended_item_id])`

which is executed when a provider uses the API to add a 'date attended'
to a fee.

This code looks up the fee record by `uuid`, which is an unindexed field
in the  `fees` table. In production this table contains over 2.4 million
records, and this query is taking increasingly long to run.

The speed of query can be improved by adding an index to the `uuid` field.