Unpin alpine gmp library
Due to a security vulnerability in an earlier version of the ruby alpine
base docker impage, the gmp library was pinned to a specifc version.

https://github.com/ministryofjustice/Claim-for-Crown-Court-Defence/commit/6a046435eff016910113e0ce0986a5eb27c27382

This vulnerability has now been resolved and so gmp can be removed from
the libraries installed in the docker image. This will result in the
version installed in the base image being used.