Fix XSS javascript vulnerability
`app/webpack/javascripts/modules/Modules.AllocationDataTable.js` uses the
`.html` jquery method to build a message to display when a caseworker
admin allocates claims to a caseworker.

Snyk has identified this as a cross-site scripting vulnerability:
```
Unsanitized input from data from a remote resource flows into html,
where it is used to dynamically construct the HTML page on client side.
This may result in a DOM Based Cross-Site Scripting attack (DOMXSS)
```

This is fixed by replacing the `.html` method with `.text` which escapes
any html passed into it, removing the vulnerability while not affecting
the message displayed to users.