Fix DataTables pagination styling
Upgrading to v2 of datatables breaks existing styling of the pagination
funtionality on the Allocations page.

This commit fixes the styling by configuring the text of the Previous and
Next buttons, and adding bespoke styling to set the font family and size,
and button to ensure that the appearance remains as before.