Fix failing test
Version 3.13.2 of the rspec-expectations gems updates the `include` matcher to
raise an exception if no arguments are passed. See https://github.com/rspec/rspec-expectations/pull/1479

The `expected_interim_fees_for` helper is meant to return an array of expected
options in a case type selection to be used in the `include` matcher. However,
in some cases `nil` is returned resulting in no arguments resulting in an
exception. This was masking failures as the test was checking that the list
included items in an empty array and was therefore always passing.

It turns out that an update to `govuk-form-builder` from some time ago changed
the id of select boxes and so the list of options being tested was always
empty. This should have immediately resulted in a a failure but it was missed
due to the bug in `expected_interim_fees_for`.