Rubocop: Fix Rails/RootPathnameMethods
Version 2.26.0 of `rubocop-rails` adds support for `Dir.[]` in the
Rails/RootPathnameMethods cop. It now replaces `Dir[Rail.root.join(...)]`
with `Rails.root.glob(...)`. This is not exactly identical as formerly an array
of string is generated whilst the new version creates an array of `Pathname`s.
In the context in which this has been changed the behaviour is equivalent.