Silence CSP alerts for chrome-extensions
Since introducing a Content Security Policy we have seen a large
volume of alerts for CSP violations caused by Chrome extensions
(indicated by a `source-file` of `chrome-extension`). This results in a
lot of noise in the laa-cccd-alerts Slack channel which distracts from
other alerts.

We have no control over users' browser extensions. To reduce the impact
on that channel, this commit updates the CspReportsController so that
Slack notifications are not created and sent when violations are caused
by Chrome extensions.