Update Google Analytics javascript
Updates the code in app/views/layouts/_analytics.js.haml to bring it in
line with Google Analytics 4.

The tags used by GA4 can be used by both Analytics and Google Tag
Manager. This commit removes the existing javascript code for each of
those services and replaces it with a single snippet which functions
for both.

As a result, the Google Tag Manager ids in `app-config.yaml` can both
be removed.