Fix flakey test
When a claim is created from a factory there is a defendant automatically
created with a random name. Occasionally this causes the tests that search for
claims by defendant name to fail as the random name matches the name used in
the tests.

This is fixed by using the `create_defendant_and_rep_order: false` option.