Log request data in response
When requests are sent to the API using JSON in the body then the data is not
available to be logged before the requests is processed. All parameters are
added to api.request.input during the processing of the request and this is
available for logging after. Therefore, to ensure that the correct data is
being logged it is added to the log after the request as well as before.