Reintegratie Google Tag Manager
The use of the Google Tag Manager API (as opposed to Google Analytics)
was removed in a previous commit as it did not seem to be needed.
However in order to make best use of functionality such as key events
it seems like we might still need to use GTM. This commit reinstates
the functionality.