Use Pagy on caseworker claims pages
Replaces the uses of the Kaminari gem for pagination with Pagy on the
caseworker claims pages.

This is more complex than the change to the provider claims pages due to
the way claims data is fetched from the API. Pagination is currently
done by the API, so it is necessary to make changes there too.

This commit

* updates `API::V2::CaseWorkers::Claim` to use Pagy to paginate claims
and pass the resulting Pagy object to `API::Entities::PaginatedCollection`
* updates `API::Entities::PaginatedCollection` to read pagination data
from the Pagy object and include it in the API response
* updates `CaseWorkers::ClaimsController` to create a new Pagy object on
the client side based on pagination data received from the API
* updates views to use the Pagy object when displaying pagination controls