Use Pagy on caseworker admin allocation pages
Replaces the uses of the Kaminari gem for pagination with Pagy on the
caseworker allocation pages.

This is more complex than the change to the provider claims pages due to
the way claims data is fetched from the API. Pagination is currently
done by the API, so it is necessary to make changes there too.

Similar to the previous commit it is necessary to create a new Pagy object
in the CaseWorkers::Admin::AllocationsController based on pagination data
received from the API. This also fixes a bug(?) where the `sort_and_paginate`
method was never being called in that controller.