Update VCR cassettes
Requests to the API have changed as Pagy indexes pages from 1, not 0 used
by Kaminari. This means the `pages` parameter in API calls to the
`caseworker/claims` endpoint now use 1 instead of 0, and VCR cassettes
need to be updated to reflect this.

This also updates a feature test which had become inaccurate to a
change of onscreen wording as a result of the use of the @pagy_info
helper method to display pagination data.