Allow multiple attachments to be added to messages
The multi file upload component of from the MoJ Fronend library is used to
enable users to attach multiple files to messages on claims. The component
requires two endpoints on the backend to upload and delete files. Ideally these
would be accessed using POST and DELETE actions but the component expects to
use POST with both.

The response from the update request is expected to contain an `originalname`
and a `filename` while the delete request sends the filename to delete. This
does not fit well with how CCCD identifies attachments and so the
`originalname` is used for the actual name of the file and the `filename` is
used for the database id.

The module `Modules.MultiFileUpload.js` is a wrapper for the multi file upload
component.