CTSKF-1009 Rename existing injection warnings
Warnings are displayed when certain fees are not injected into CCR. The
wording of some of these warnings could be better.

This changes "CLAR fees not injected" to "Paper heavy case or unused materials
fees not injected" and "CAVs not injected" to "Conference fees not injected"
in order to remove acronyms and improve clarity.

As part of this change, the width of the warning class is increased to
allow for the more verbose message.