Reformat injection messages on Allocation page
Updates the appearance of error and warning messages on claims on the
Caseworker Allocation page. Instead of being randomly scattered around
the table row, they are now grouped under the Case Number in GDS-like
format.

This includes errors as a result of injection failures, and warnings
about non-injected fees and disk evidence.