Refactor param handling
With Rails 7.1 a callback will raise an exception if the action doesn't exist.
This causes a problem for a mixin that is used in controllers where not all
actions may be defined.

In `PasswordHelpers` the callbacks are used to set strong parameter, which
does not normally need to be set as an instance variable. A method with
memoization works well in this case.