Update MAAT reference matching
The MAAT reference from Court Data was being taken from the
'representation_order' field of the defendant. This is not always present and
it should instead come from the 'laa_reference' option of the offence
summaries. There may now be multiple references and these are all matched
against all the references that the provider lists, not just the earliest.