Implement skipped jasmine tests
Four jasmine tests in Modules.MangementInformation_spec.js have never
been defined and are marked as TODO. This commit implements the tests
so that we have test coverage for all of our javascript.

* adds tests for the `blockDisabledLinkClick`, `dateInputEvent`,
`disableDownloadButton` and `activateDownload` functions.
* this highlights an issue with the `disableDownloadButton` function in
the `ManagementInformation` module, which, when the disable button is
clicked, adds sets `aria-disabled` to be `true` and then immediately
sets it to be `false` again. This removes the instruction to set it to
`false`.
* it also reveals a problem in `modules-amountAssessed_spec.js` where
a test overrides the jQuery method `$.fn.addClass` without resetting it,
which then pollutes the use of that method in other tests, leading to
intermittent failures. This commit replaces that overriding with a spy.
* fixes some minor code quality and indentation issues