Remove 'disabled' class from download button
The provisional_assessments_date_download button on the MI screen is
disabled when the data selector does not contain a valid date. This is
done by adding a `disabled` class to the element.

The `govuk-button` class automatically changes the appearance of the
button depending on the presence of the `disabled` attribute without
needing the extra class. This commit removes the unnecessary class.

As a result it also updates the `blockDisabledLinkClick` function to
disable the button based on the presence of the `disabled` attribute.