CTSKF-1108 Display full defendant name
When viewing the claim summary page as a caseworker, the defendant name
is currently split over two lines as first name and last name.

This is inconvenient for users, who need to copy and paste the full name,
so this commit replaces the two lines with a single line displaying the
full name.