Add beta testing toggle
Beta testing is enabled by creating a copy of a particular partial with
the name appended with `_beta`. The `beta_test_partial` can be used in a
`render` to select the correct partial to use depending on whether the
beta view is enabled.

The beta view is enabled depending on the session cookie `beta_testing`.
By default, the beta view is enabled and it can be turned on or off by
visiting `/beta/enable` and `/beta/disable`, which will redirect back to
the previous page.