Improve labelling of VAT fields on determination table
When assessing an AGFS claim, a caseworker sees three fields relating to
VAT:

* The value of the claim, before VAT - labelled as "Total excluding VAT"
* The VAT value for the claim - labelled as "Total (including 20% VAT)"
* The total value for the claim - labelled as "Total inc VAT rate:"

The second line makes no sense, as this displays the VAT value not the total.
The third line could be worded better, ideally using the current wording
from the second line.

This commit updates the relevant translation file to use the correct
wordings in the correct places.