Rubocop: Fix RSpec/MultipleExpectations
Some of these multiple expectations specs are testing that certain details are
or are not copied from the original to the clone and there are tests for the
original claim that are not strictly necessary. These could be refactored such
that the set up ensures that the original claim is correct.