Fix 'no method error' error
The defendant uplift validations raises an exception when the quantity on a fee
is `nil` as these quantities are summed. This was probably masked prior to
Rails 7.1 as the Rails version of the `sum` method ignored blank values in.
In Rails 7.1 the Rails version of `sum` is removed in favour of the Ruby
version and this fails if the array does not contain only numbers.

It is unclear how the quantity on a fee can be set to `nil`.