Allow claims to be created for contingency MAAT ID
Special legislation is being laid to come into force on 27th June 2025,
in response to enabling ongoing legal aid contingency work.

As part of these legislation changes providers will be able to self grant
representation orders against a dummy MAAT ID 900900.

We need to allow providers to submit claims to CCCD against ID 900900.
To do this, this commit relaxes the validation of representation order
IDs to allow that specific ID to be entered in addition to all usual
valid IDs.

This is done by updating the `maat_regexp` regular expression in
`config/settings.yml`.