Ignore brakeman warning about Rails 7.1.5.1
Brakeman is failing due to a new warning about the end of Rails 7.1.5.1
support on 2025-10-01. We have plans to upgrade to 7.2 before then, so
all this warning is achieving is blocking our deployment pipeline.

This adds the warning to the brakeman.ignore file so that we can
continue to deploy CCCD.