Set AWS_REGION for assets:precompile
A change in Rails means that S3 configuration is loaded during the
assets:precompile task. When this runs in the Dockerfile the
AWS_REGION environment variable is not set so this fails. As a
work-around this variable is set for this command. It is possible that
this can be removed when upgrading to Rails 8.0.

See: https://github.com/rails/rails/issues/54139