CTSKF-1466 Update ModSecurity rules
In CCCD, users are experiencing occasional failures when trying to send
messages. There is no error but the message fails to send.

This is a result of false positives identified by ModSecurity rules
932370 and 932380. These rules identify two instances of Remote Command
Execution: Windows Command Injection and are matching messages that
contain punctuation such as colons, slashes, semicolons; command-like
words (e.g., 'start', 'statement', and 'response'); and dates and
monetary values.

This change adds 932370 and 932380 to the list of disabled ModSecurity rules
on the /messages page.

This will prevent valid messages being blocked as false positives while
other parts of the application will continue to be protected by those rules.