Hide warrant fee details if no fee claimed
Migrating to the component gem for summary lists has led to a small bug
on the provider summary page where a blank section for warrant fee details
is displayed even if no warrant fee has been claimed.

This occurs because a component which was previously within a
conditional is now not. This commit fixes the indentation of that list
so that it is now only displayed if `info.warrant_fee_paid?` is true.