Use GOVUK Components gem for pagination styling
Implements the `govuk-components` gem for styling pagination navigation.

This is done by updating the `app/views/shared/_pagination.html.haml`
partial to call the `govuk_pagination` method, and updating
`app/webpack/stylesheets/_navigation.scss` to ensure the navigation
continues to be displayed in the same place.

This ensures the application is consistent with GDS designs and improves
accessibility and usability.